削除時ダイアログを表示

+++ twitter_clone/src/resources/js/components/ProfileTweet.vue
@@ -149,6 +149,7 @@ export default {
         },
 
         removeTweet(id) {
+            if(confirm('削除してよろしいですか?'))
             axios.delete("/api/deleteTweet/" + id
             ).then((res) => {
                 this.timeLines = this.timeLines.filter(item => item.id !== id)
